trackgen: fix gcc warning
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@9374 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 0377f0509cee7bb22aa8b796a8c2b483d36d176d
Former-commit-id: fc4bb1e9deb532a342a0693987b430265dcb178c

+++ src/tools/trackgen/ac3d.cpp
@@ -912,7 +912,7 @@ void Ac3d::Object::removeUnusedVertices()
         }
     }
     int originalIndex = 0;
-    for (int i = 0; i < vertices.size(); originalIndex++)
+    for (size_t i = 0; i < vertices.size(); originalIndex++)
     {
         if (usedIndex.find(originalIndex) == usedIndex.end())
         {
